+++ delivery-service/src/main/java/com/pickple/delivery/application/dto/response/DeliveryStartResponseDto.java
@@ -1,4 +1,4 @@
-package com.pickple.delivery.application.dto;
+package com.pickple.delivery.application.dto.response;
 
 import java.util.UUID;
 import lombok.Builder;
