+++ delivery-service/src/main/java/com/pickple/delivery/application/mapper/DeliveryDetailMapper.java
@@ -1,10 +1,10 @@
 package com.pickple.delivery.application.mapper;
 
-import com.pickple.delivery.application.dto.DeliveryDetailCreateRequestDto;
-import com.pickple.delivery.application.dto.DeliveryDetailCreateResponseDto;
+import com.pickple.delivery.application.dto.request.DeliveryDetailCreateRequestDto;
+import com.pickple.delivery.application.dto.response.DeliveryDetailCreateResponseDto;
 import com.pickple.delivery.application.dto.DeliveryDetailInfoDto;
 import com.pickple.delivery.domain.model.DeliveryDetail;
-import com.pickple.delivery.domain.repository.DeliveryDetailInfoProjection;
+import com.pickple.delivery.domain.repository.projection.DeliveryDetailInfoProjection;
 import com.pickple.delivery.presentation.request.DeliveryDetailCreateRequest;
 import java.util.UUID;
