+++ delivery-service/src/main/java/com/pickple/delivery/application/service/DeliveryApplicationService.java
@@ -3,18 +3,18 @@ package com.pickple.delivery.application.service;
 import com.pickple.common_module.exception.CommonErrorCode;
 import com.pickple.common_module.exception.CustomException;
 import com.pickple.common_module.infrastructure.messaging.EventSerializer;
-import com.pickple.delivery.application.dto.DeliveryCreateRequestDto;
+import com.pickple.delivery.application.dto.request.DeliveryCreateRequestDto;
 import com.pickple.delivery.application.dto.DeliveryDetailInfoDto;
 import com.pickple.delivery.application.dto.DeliveryInfoDto;
-import com.pickple.delivery.application.dto.DeliveryInfoResponseDto;
+import com.pickple.delivery.application.dto.response.DeliveryInfoResponseDto;
 import com.pickple.delivery.application.events.DeliveryCreateFailureEvent;
 import com.pickple.delivery.application.events.DeliveryCreateResponseEvent;
-import com.pickple.delivery.application.dto.DeliveryStartRequestDto;
-import com.pickple.delivery.application.dto.DeliveryStartResponseDto;
+import com.pickple.delivery.application.dto.request.DeliveryStartRequestDto;
+import com.pickple.delivery.application.dto.response.DeliveryStartResponseDto;
 import com.pickple.delivery.application.mapper.DeliveryMapper;
 import com.pickple.delivery.domain.model.enums.DeliveryCarrier;
 import com.pickple.delivery.domain.model.enums.DeliveryType;
-import com.pickple.delivery.domain.repository.DeliveryInfoProjection;
+import com.pickple.delivery.domain.repository.projection.DeliveryInfoProjection;
 import com.pickple.delivery.domain.repository.DeliveryRepository;
 import com.pickple.delivery.domain.model.Delivery;
 import com.pickple.delivery.exception.DeliveryErrorCode;
