+++ delivery-service/src/main/java/com/pickple/delivery/domain/model/DeliveryDetail.java
@@ -1,6 +1,6 @@
 package com.pickple.delivery.domain.model;
 
-import com.pickple.delivery.application.dto.DeliveryDetailCreateRequestDto;
+import com.pickple.delivery.application.dto.request.DeliveryDetailCreateRequestDto;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
