+++ delivery-service/src/main/java/com/pickple/delivery/domain/repository/projection/DeliveryInfoProjection.java
@@ -1,4 +1,4 @@
-package com.pickple.delivery.domain.repository;
+package com.pickple.delivery.domain.repository.projection;
 
 import com.pickple.delivery.domain.model.enums.DeliveryStatus;
 import com.pickple.delivery.domain.model.enums.DeliveryType;
